LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/cssom-view/scrollIntoView-stuck.tentative-expected.txt



FAIL CSSOM View - scrollIntoView doesn't consider scroll-padding when target is stuck assert_equals: Shouldn't have scrolled expected 2324 but got 2422
